refactor(review): name the note byte check for size, not bounds
`noteBounds.ts` measures a note's serialized *bytes*, but CLAUDE.md reserves
"bounds" for one concrete visible extent, and the codebase honors that
spatially everywhere else (`rowBounds`, `hunkBounds`, `selectedNoteBounds` in
DiffPane, `VerticalBounds`). This was the one vocabulary break, and it collided
greppably with the spatial note-card concept in the same feature area.

Rename the module to `noteSize.ts` and `reviewNoteWithinBounds` to
`reviewNoteWithinSizeLimit`; `MAX_REVIEW_NOTE_BYTES` and `reviewNoteByteLength`
were already unambiguous and stay put. The conformance fixture file and its
exports follow, along with the audit doc's references to the current code.
