+++ test/review-conformance/noteSize.ts
@@ -11,16 +11,16 @@
  * Sizes are stated relative to the shared bound rather than as literals, so the corpus
  * still means the same thing if the bound moves.
  */
-import { MAX_REVIEW_NOTE_BYTES } from "../../src/core/review/noteBounds";
+import { MAX_REVIEW_NOTE_BYTES } from "../../src/core/review/noteSize";
 import type { ReviewNoteV1 } from "../../src/core/review/types";
 
-export interface ReviewNoteBoundsFixture {
+export interface ReviewNoteSizeFixture {
   id: string;
   /** What makes this note adversarial, in one line. */
   description: string;
   build: () => ReviewNoteV1;
   /** Hand-written from the semantics — never captured from the measurement. */
-  withinBounds: boolean;
+  withinSizeLimit: boolean;
 }
 
 /** One minimal note with the given text fields; everything else is framing. */
@@ -42,24 +42,24 @@ const FRAMING_BYTES = JSON.stringify(note({})).length;
 /** ASCII filler of an exact byte length. */
 const filler = (bytes: number) => "x".repeat(Math.max(0, bytes));
 
-export const REVIEW_NOTE_BOUNDS_FIXTURES: readonly ReviewNoteBoundsFixture[] = [
+export const REVIEW_NOTE_SIZE_FIXTURES: readonly ReviewNoteSizeFixture[] = [
   {
     id: "empty-note",
     description: "Framing alone is far below the bound.",
     build: () => note({}),
-    withinBounds: true,
+    withinSizeLimit: true,
   },
   {
     id: "whole-note-exactly-at-the-bound",
     description: "Summary sized so the serialized note lands on the limit exactly.",
     build: () => note({ summary: filler(MAX_REVIEW_NOTE_BYTES - FRAMING_BYTES) }),
-    withinBounds: true,
+    withinSizeLimit: true,
   },
   {
     id: "whole-note-one-byte-over",
     description: "The same note plus one byte: over the limit as a whole.",
     build: () => note({ summary: filler(MAX_REVIEW_NOTE_BYTES - FRAMING_BYTES + 1) }),
-    withinBounds: false,
+    withinSizeLimit: false,
   },
   {
     id: "every-field-fits-but-the-note-does-not",
@@ -71,20 +71,20 @@ export const REVIEW_NOTE_BOUNDS_FIXTURES: readonly ReviewNoteBoundsFixture[] = [
         rationale: filler(MAX_REVIEW_NOTE_BYTES - 1),
         markup: filler(MAX_REVIEW_NOTE_BYTES - 1),
       }),
-    withinBounds: false,
+    withinSizeLimit: false,
   },
   {
     id: "multibyte-summary-under-the-per-character-limit",
     description:
       "A summary of four-byte characters that is well under the bound counted as characters and over it counted as bytes.",
     build: () => note({ summary: "🧪".repeat(MAX_REVIEW_NOTE_BYTES / 4) }),
-    withinBounds: false,
+    withinSizeLimit: false,
   },
   {
     id: "multibyte-summary-just-inside",
     description: "The same characters, one short of filling the bound with framing included.",
     build: () =>
       note({ summary: "🧪".repeat(Math.floor((MAX_REVIEW_NOTE_BYTES - FRAMING_BYTES) / 4)) }),
-    withinBounds: true,
+    withinSizeLimit: true,
   },
 ];
